Proposal: Send battery level to JS
I know you are working on a JS API, and I'm not expecting this to be merged. I'm simply proposing that receiving battery level information is part of this API, and this PR is one way of fetching this.

To read the battery-level, the used web page need to define a `window.SEB = {}`, and then poll window.SEB.battery with a reasonable interval to see if battery data has been received.